Use only systemd arch identifiers for user parameters
The difference in architecture identifiers from upstream binary releases
and what systemd expects in the extension metadata or the update
manifest easily causes confusion.
Recommend the use of the systemd identifiers as they will show up in
filenames. Map the identifiers internally and try to do the right thing
even if a slighly different identifier was called.