kubernetes: Place cni binaries in /opt/cni/bin
Containerd expects cni binaries in /opt/cni/bin. We need to make sure that
directory exists before running the cp, and need to run cp in a way that
doesn't result in the source directory being created in target (we only want
the executables).

The /opt/libexec thing was an experiment that can be removed as well, as we
found another way of dealing with volume plugins.

Signed-off-by: Jeremi Piotrowski <jpiotrowski@microsoft.com>