release action: Create new release from list, include old release
This change updates the release GitHub Actions workflow to build a new
release instead of modifying an existing one. The previous "latest"
release will be deleted.

The release will contain all sysexts listed in
"release_build_versions.txt" as well as all sysexts included in the
previous latest release. It will generate a Release.md to be included in
the release.

The change also updates the workflow to only run when the tag "latest"
is pushed.

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>