flix and flatwrap: Add tools to convert a chroot to a sysext
Existing binaries can be bundled in a sysext image if they use their
own library path or do what Flatpak does and run in a chroot. This
introduces the Flix tool which creates sysext images from a given chroot
and a set of binaries in it, and then patches the ELF rpath section and
to refer to a custom library path.
The other tool introduced here is the Flatwrap tool that creates sysext
images from a given chroot and a set of binaries in it, and sets up
wrapper entry points with bwrap/unshare into the chroot.
This also introduces a helper oci-rootfs.sh to create a useable chroot
from a container image.