release.yaml: Split artifact upload and don't delete release
Artifact upload is buffered in memory, which is a limitation of the action. We
need to chunk the upload manually to work around this - by splitting it into
groups that fit within the memory limit of the worker.

We also don't need to delete the release - something that has caused us trouble
when the action fails during upload - and instead update the payload of
existing releases.  We can either keep pushing the "latest" tag (like now) or
push date base release tags and rename to "latest" manually.