release_dispatcher.sh: fix version array
The version variable is used as both an array and a scalar type in
release_dispatcher.sh. When being set to a scalar value and previously
being used as array, only the first array member is overwritten. This
leads to erroneous build dispatches because versions are mixed up.

We explicitly un-set the variable at the end of the version loop

We also check the presence of <extension>:<version> pairs in the build
array before adding new ones to prevent cases where "latest" and a
pinned version are identical

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>