GitHub release builds: build log in files, add dates
This change re-adds build logs to releases but uses separate files
instead of adding the log to Release.md.

This addresses a corner case where Release.md grew too large to be used
as release text (see
https://github.com/flatcar/sysext-bakery/pull/138#issuecomment-2771519319).

Also, the PR adds build timestamps to release texts as well as removes
the contents of SHA256SUMS from the SHA256SUMS metadata Release.md to
avoid potential issues in the future as our release count grows. The
SHA256SUMS file is part of that release, no need to repeat it in
Release.md

Signed-off-by: Thilo Fromm <thilofromm@microsoft.com>